Address metrics boundary and cardinality
