+++ docs/qa/v3.1.md
@@ -207,8 +207,8 @@ is attached.
 - [ ] **v3.1.0 release blocker:** Process/runtime health metrics are available from safe
       server-side collection.
 - [x] **implemented source behavior:** dChat request totals and latency are labeled only by bounded
-      provider and outcome categories, including browser chat reports forwarded to the server
-      metrics registry.
+      provider and outcome categories, including browser chat reports forwarded through validated
+      same-origin `POST /metrics` payloads to the server metrics registry.
 - [x] **implemented source behavior:** token.place dependency metrics distinguish successes, timeouts,
       rate limits, malformed responses, and server failures.
 - [x] **implemented source behavior:** Provider-selection failures and fallback behavior are measurable
